Moved time related functions from ScheduleDeck to TimeService
