+++ src/opm/parser/eclipse/EclipseState/Schedule/ScheduleDeck.cpp
@@ -27,60 +27,10 @@
 #include <opm/common/OpmLog/OpmLog.hpp>
 #include <opm/common/utility/OpmInputError.hpp>
 #include <opm/common/utility/String.hpp>
+#include <opm/common/utility/TimeService.hpp>
 
 namespace Opm {
 
-
-namespace {
-
-std::time_t mkdatetime(int in_year, int in_month, int in_day, int hour, int minute, int second) {
-    const auto tp = TimeStampUTC{ TimeStampUTC::YMD { in_year, in_month, in_day } }
-        .hour(hour).minutes(minute).seconds(second);
-
-    std::time_t t = asTimeT(tp);
-    {
-        /*
-          The underlying mktime( ) function will happily wrap
-          around dates like January 33, this function will check
-          that no such wrap-around has taken place.
-        */
-        const auto check = TimeStampUTC{ t };
-        if ((in_day != check.day()) || (in_month != check.month()) || (in_year != check.year()))
-            throw std::invalid_argument("Invalid input arguments for date.");
-    }
-    return t;
-}
-
-std::time_t mkdate(int in_year, int in_month, int in_day) {
-    return mkdatetime(in_year , in_month , in_day, 0,0,0);
-}
-
-std::time_t timeFromEclipse(const DeckRecord &dateRecord) {
-    const auto &dayItem = dateRecord.getItem(0);
-    const auto &monthItem = dateRecord.getItem(1);
-    const auto &yearItem = dateRecord.getItem(2);
-    const auto &timeItem = dateRecord.getItem(3);
-
-    int hour = 0, min = 0, second = 0;
-    if (timeItem.hasValue(0)) {
-        if (sscanf(timeItem.get<std::string>(0).c_str(), "%d:%d:%d" , &hour,&min,&second) != 3) {
-            hour = min = second = 0;
-        }
-    }
-
-    // Accept lower- and mixed-case month names.
-    std::string monthname = uppercase(monthItem.get<std::string>(0));
-
-    std::time_t date = mkdatetime(yearItem.get<int>(0),
-                                  TimeService::eclipseMonthIndices().at(monthname),
-                                  dayItem.get<int>(0),
-                                  hour,
-                                  min,
-                                  second);
-    return date;
-}
-
-}
 
 
 ScheduleBlock::ScheduleBlock(const KeywordLocation& location, ScheduleTimeType time_type, const time_point& start_time) :
@@ -188,13 +138,13 @@ ScheduleDeck::ScheduleDeck(const Deck& deck, const ScheduleRestartInfo& rst_info
         // Use the 'START' keyword to find out the start date (if the
         // keyword was specified)
         const auto& keyword = deck.getKeyword("START");
-        start_time = TimeService::from_time_t( timeFromEclipse(keyword.getRecord(0)) );
+        start_time = TimeService::from_time_t( TimeService::timeFromEclipse(keyword.getRecord(0)) );
     } else {
         // The default start date is not specified in the Eclipse
         // reference manual. We hence just assume it is same as for
         // the START keyword for Eclipse E100, i.e., January 1st,
         // 1983...
-        start_time = TimeService::from_time_t( mkdate(1983, 1, 1) );
+        start_time = TimeService::from_time_t( TimeService::mkdate(1983, 1, 1) );
     }
 
     this->m_restart_time = TimeService::from_time_t(rst_info.time);
@@ -223,7 +173,7 @@ ScheduleDeck::ScheduleDeck(const Deck& deck, const ScheduleRestartInfo& rst_info
                 std::time_t nextTime;
 
                 try {
-                    nextTime = timeFromEclipse(record);
+                    nextTime = TimeService::timeFromEclipse(record);
                 } catch (const std::exception& e) {
                     const OpmInputError opm_error { e, keyword.location() } ;
                     OpmLog::error(opm_error.what());
